Collapsed field and ring definitions into one module.
