+++ src/kem/kyber768/serialize.rs
@@ -1,6 +1,6 @@
 use crate::kem::kyber768::{
     parameters::{BYTES_PER_RING_ELEMENT, COEFFICIENTS_IN_RING_ELEMENT},
-    ring::KyberPolynomialRingElement,
+    arithmetic::KyberPolynomialRingElement,
 };
 
 /// This file contains instantiations of the functions
